fix: Fix Circular Dependency
