+++ src/applications/disability-benefits/all-claims/content/form0781/workflowChoicePage.jsx
@@ -15,6 +15,12 @@ import { form0781HeadingTag, titleWithTag } from '../form0781';
 // eslint-disable-next-line import/no-cycle
 import { checkValidations } from '../../utils/submit';
 
+import {
+  form0781WorkflowChoiceDescription,
+  form0781WorkflowChoices,
+  form0781WorkflowChoiceLabels,
+} from './workflowChoices';
+
 export const workflowChoicePageTitle =
   'Option to add a statement in support of mental health conditions';
 
@@ -94,24 +100,6 @@ export const workflowChoicePageDescription = formData => {
     </>
   );
 };
-
-export const form0781WorkflowChoiceDescription =
-  'Do you want to add a statement in support of mental health conditions?';
-
-export const form0781WorkflowChoices = {
-  COMPLETE_ONLINE_FORM: 'optForOnlineForm0781',
-  SUBMIT_PAPER_FORM: 'optForPaperForm0781Upload',
-  OPT_OUT_OF_FORM0781: 'optOutOfForm0781',
-};
-
-export const form0781WorkflowChoiceLabels = Object.freeze({
-  [form0781WorkflowChoices.COMPLETE_ONLINE_FORM]:
-    'Yes, and I want to answer the questions online.',
-  [form0781WorkflowChoices.SUBMIT_PAPER_FORM]:
-    'Yes, and I want to fill out a PDF to upload.',
-  [form0781WorkflowChoices.OPT_OUT_OF_FORM0781]:
-    'No, I don’t want to add this form to my claim.',
-});
 
 export const traumaticEventsExamples = (
   <va-accordion open-single>
