+++ src/applications/disability-benefits/all-claims/content/form0781/workflowChoices.jsx
@@ -1,0 +1,17 @@
+export const form0781WorkflowChoiceDescription =
+  'Do you want to add a statement in support of mental health conditions?';
+
+export const form0781WorkflowChoices = {
+  COMPLETE_ONLINE_FORM: 'optForOnlineForm0781',
+  SUBMIT_PAPER_FORM: 'optForPaperForm0781Upload',
+  OPT_OUT_OF_FORM0781: 'optOutOfForm0781',
+};
+
+export const form0781WorkflowChoiceLabels = Object.freeze({
+  [form0781WorkflowChoices.COMPLETE_ONLINE_FORM]:
+    'Yes, and I want to answer the questions online.',
+  [form0781WorkflowChoices.SUBMIT_PAPER_FORM]:
+    'Yes, and I want to fill out a PDF to upload.',
+  [form0781WorkflowChoices.OPT_OUT_OF_FORM0781]:
+    'No, I don’t want to add this form to my claim.',
+});
